Refactor sidebar navigation and test form components
